[refactor] : 오르막수 반환값 변경

+++ src/test/java/me/boj/dp/ps_11057/AscendingTest.java
@@ -8,9 +8,9 @@ class AscendingTest {
 
     @Test
     void solve() {
-        int r1 = Ascending.solve(1);
-        int r2 = Ascending.solve(2);
-        int r3 = Ascending.solve(3);
+        long r1 = Ascending.solve(1);
+        long r2 = Ascending.solve(2);
+        long r3 = Ascending.solve(3);
 
         assertThat(r1).isEqualTo(10);
         assertThat(r2).isEqualTo(55);
